FIX: optimize: use int return code instead of intptr_t

+++ scipy/optimize/rectangular_lsap/rectangular_lsap.h
@@ -37,8 +37,8 @@ extern "C" {
 
 #include <stdint.h>
 
-intptr_t solve_rectangular_linear_sum_assignment(intptr_t nr, intptr_t nc, double* input_cost,
-                                                 int64_t* col4row);
+int solve_rectangular_linear_sum_assignment(intptr_t nr, intptr_t nc, double* input_cost,
+                                            int64_t* col4row);
 
 #ifdef __cplusplus
 }
